+++ app/api/email-check/route.ts
