+++ app/api/join/route.ts
